requirements.txt
# Core dependencies
numpy>=1.24.0,<2.0.0
gradio>=4.40.0,<5.0.0
streamlit>=1.28.0

# Environment variables
python-dotenv>=1.0.0

# Audio processing (requires portaudio system package)
PyAudio>=0.2.14

# Video processing (requires ffmpeg installed on system)
moviepy>=2.0.0

# Transcription engines
faster-whisper>=1.0.0
whisperx>=3.0.0  # Optional: Advanced transcription with diarization
openai>=1.0.0

# Audio Analysis (Tonal/Prosodic features)
librosa>=0.10.0

# Text Analysis
transformers>=4.30.0
torch>=2.0.0
torchvision>=0.15.0
pillow>=9.0.0
easyocr>=1.7.0
opencv-python>=4.8.0

# Note: WhisperX requires additional system dependencies:
# - CUDA 12.8 (for GPU support)
# - ffmpeg
# Install: pip install whisperx
# For diarization, accept HF model agreements and set HF_TOKEN

# Note: System requirements
# - macOS: brew install portaudio ffmpeg
# - Ubuntu/Debian: sudo apt-get install portaudio19-dev ffmpeg
# - Windows: Install ffmpeg from https://ffmpeg.org/download.html
